Added install location to AndroidManifest.xml
android:installLocation="auto"

Without this line the game is forced to be installed in the internal and often scarce phone memory. Neither can it be moved to the sdcard through the system settings.

With this line the game may be installed on the external storage. The system will decide where to install the game based on several factors. The user can also move the game between the two locations.